chore(archai): Creates a cv package and moves cifar10_models and networks to it.
