+++ archai/cv/cifar10_models/googlenet.py
