+++ archai/cv/networks/shakeshake/shake_resnet.py
