+++ archai/cv/networks/shakeshake/shake_resnext.py
