+++ archai/cv/networks/shakeshake/shakeshake.py
